Restructure to make the implementation of wave plus functionality cleanly separated
