+++ waveplus_test.go
